add: Checkin Dialog
-Made dialog name more descriptive.
-Organized some files into packages
-Did some string extracting into `strings.xml`.
-Checkin now links to the Checkin dialog instead of  Checkout linking do the Checkin dialog.
-Minor tidying with constructor suppression.
