+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="checked_in">Checked In</string>
     <string name="check_out">Check Out</string>
     <string name="check_in">Check In</string>
+    <string name="date_due_detail">Due %s</string>
 </resources>
